move FontInstance out of webrender_api so it is easier to extend
